requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Embeddings & Vector DB
sentence-transformers>=2.2.0
chromadb>=0.4.22
torch>=2.1.0

# File Watching
watchdog>=3.0.0

# Markdown Processing
python-frontmatter>=1.1.0
tiktoken>=0.5.0
markdownify>=0.11.0  # HTML to Markdown conversion
beautifulsoup4>=4.12.0  # HTML parsing for GDoc CSS stripping

# Configuration
pyyaml>=6.0

# Claude API
anthropic>=0.18.0

# Google APIs
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-xdist>=3.5.0  # Parallel test execution
pytest-playwright>=0.7.0  # Browser-based UI tests
playwright>=1.40.0
httpx>=0.26.0

# Scheduling
croniter>=2.0.0

# Utilities
python-dotenv>=1.0.0
rapidfuzz>=3.6.0  # Fast fuzzy string matching for entity resolution
python-multipart
psutil

# macOS Integrations (optional)
pyobjc-framework-Contacts>=10.0; sys_platform == "darwin"  # Apple Contacts integration
